fix bug

+++ processing/src/main/java/org/apache/druid/query/UsageUtils.java
@@ -125,7 +125,7 @@ public class UsageUtils
     }
 
     if (value instanceof Number) {
-      return (((Number) value).intValue()) == 0;
+      return (((Number) value).doubleValue()) == 0;
     }
     return false;
   }
